+++ app/components/atoms/paragraph/paragraph.stories.tsx
@@ -1,6 +1,6 @@
 import {ComponentStory, ComponentMeta} from '@storybook/react';
 
-import {Paragraph} from './Paragraph.component';
+import {Paragraph} from './paragraph.component';
 
 export default {
   title: 'Atoms/Paragraph',
@@ -11,11 +11,11 @@ const Template: ComponentStory<typeof Paragraph> = (args) => (
   <Paragraph {...args} />
 );
 
-export const ParagraphDefault = Template.bind({});
-ParagraphDefault.args = {
-  children: 'Lorem ipsum dolor sit amet, conctetuer adipiscing elit volutpat. Hendrerit invulputatevelit esse  Utwisi enim ad minimveniam.',
+export const Default = Template.bind({});
+Default.args = {
+  children:
+    'Lorem ipsum dolor sit amet, conctetuer adipiscing elit volutpat. Hendrerit invulputatevelit esse  Utwisi enim ad minimveniam.',
   lineHeight: 1.5,
   bold: false,
   italic: false,
 };
-
